cookingbook/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
